+++ src/react-components/src/components/02-molecules/Cards/SearchResultItem.stories.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 import SearchResultItem from "./SearchResultItem";
 import { action } from "@storybook/addon-actions";
 import { withKnobs, text } from "@storybook/addon-knobs";
+import UnderlineLink from "../../01-atoms/Links/UnderlineLink";
 
 
 export default {
@@ -13,20 +14,20 @@ export default {
 export const searchResultItem = () => <SearchResultItem
   id="Result1"
   resultIndex={1}
-  totalEditions={7}
-      modifiers={[]}
-      blockName={""}
-      headingText="Investigation of un-American propaganda activities in the United States. Hear..."
-      subtitleText="Subtitle need max character count Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
-      authorUrl="#authorUrl"
-      authorName="First Last"
-      editionInfo={{
-        editionYearHeading: "2004 Edition",
-        publisherAndLocation: "Published in New York by Random House",
-        coverUrl: "https://placeimg.com/57/81/arch",
-        language: "Written in English",
-        license: "Under Creative Commons License",
-        readOnlineLink: "#read-online",
-        downloadLink: "#download"}}
-      allEditionsUrl="#allEditions"
+  modifiers={[]}
+  blockName={""}
+  headingText="Investigation of un-American propaganda activities in the United States. Hear..."
+  subtitleText="Subtitle need max character count Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+  authorLinkElement={<UnderlineLink url={"author-url"}>First Last</UnderlineLink>}
+  editionInfo={{
+    editionYearHeading: "2004 Edition",
+    publisherAndLocation: "Published in New York by Random House",
+    coverUrl: "https://placeimg.com/57/81/arch",
+    language: "Written in English",
+    license: "Under Creative Commons License",
+    readOnlineLink: "#read-online",
+    downloadLink: "#download"
+  }}
+  editionsLinkElement={<UnderlineLink url={"#allEditionsUrl"}>{`View All 7 editions`}</UnderlineLink>
+}
 />;